Improvements to the Home tab (start page)
Added a new chart for number of species per month.
Now both charts show the last 12 months from today and if there are no records hide the charts.
If there is no surveys to show, hides the map and show a message.
Better resizing of columns in the last lifers list.